fix(ios): bridge the two Kotlin number types the way Swift sees them
Both errors are the same shape — a Kotlin nullable/boxed number arriving as an
NSNumber subclass:

  RoutesScreen: `pullInbox()` returns Int, which suspend-bridges to KotlinInt,
  and `count > 0` asks NSNumber to be a BinaryInteger. Take .intValue first
  (which also drops the deprecated Int(_:) conversion).

  RouteEditorScreen: KotlinLong's `long:` initialiser is C long, i.e. Swift
  Int, not the Int64 being handed to it. `value:` is the Int64 overload, and
  matches KotlinDouble(value:) already used in TripRecorder.

+++ iosApp/Detour/RoutesScreen.swift
@@ -196,9 +196,11 @@ struct RoutesScreen: View {
     private func pullInbox() async {
         refreshing = true
         do {
-            let count = try await RouteShare.shared.pullInbox()
+            // A suspend fun returning Int arrives as KotlinInt, which is an
+            // NSNumber and so compares to nothing on its own.
+            let count = try await RouteShare.shared.pullInbox().intValue
             if count > 0 {
-                pulledCount = Int(count)
+                pulledCount = count
             }
         } catch {
             report(error)
